Add files via upload
This is the Arduino source code file for the v0.1 system. 
The source code may be readily uploaded to any Arduino ou ATMega microprocessor through the Arduino IDE.
This version supports one read channel of the ADC and one digital channel for controlling external devices. 
Expansion to more channels, as allowed by the v0.1 schematic, is simple, but a forthcoming version may already incorporate that.
In its original form, the source code does not measure the temperature and humidity. In order to incorporate that, the user needs to uncomment the "dht.readTemperature();" and subsequent lines.